Further reduced size of canvas.js
Move selection and cut logic to new file.
